+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ### Added
 - Database SQLite for storing user config.
 - Status bar menu with help, preferences and quit actions.
+- Set through link to paste of clipboard after delivering file to remote server.
+- Push notification.
+### Changed
+- Refactoring.
 
 ## [0.0.1] - 2021-07-31
 ### Added